Code_associated_Phase3/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
pandas
rapidfuzz
google-generativeai
pytest
httpx
